+++ resources/lang/de/permissions.php
@@ -14,6 +14,8 @@ return [
     'shop.barber.list.configure' => 'Frisör: Einstellungen bearbeiten',
     'people.helpers.view' => 'Helfer: Helfer betrachten',
     'people.helpers.manage' => 'Helfer: Helfer verwalten',
+    'people.helpers.casework.view' => 'Helfer: Fallarbeitsdaten anzeigen',
+    'people.helpers.casework.manage' => 'Helfer: Fallarbeitsdaten bearbeiten',
     'badges.create' => 'Badges: Badges erstellen',
     'logistics.use' => 'Logistikmodul benutzen',
     'tasks.use' => 'Aufgabenmodul benutzen',